video-minutes/requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
pyyaml>=6.0
requests>=2.31.0
jinja2>=3.1.0
python-dateutil>=2.8.0
yt-dlp>=2023.12.30
notion-client>=2.2.1
